restructure component and change maps tabs to layers
