+++ components/DataTable/index.tsx
@@ -1,0 +1,2 @@
+export { DataTable } from "./DataTable";
+export { useDataTable } from "./useDataTable";
