+++ components/Input/index.tsx
@@ -1,0 +1,3 @@
+export { InputColor } from "./InputColor";
+export { InputNumber } from "./InputNumber";
+export { InputText } from "./InputText";
