+++ features/geojson/Sidebar/Apps/ControlOptions.tsx
@@ -2,7 +2,7 @@ import { Switch } from "@mantine/core";
 import { useStore } from "@geojson/store";
 import { Options } from "@components/Options";
 
-export function MapsControlOptions() {
+export function ControlOptions() {
   const showControl = useStore(state => state.showMapControls)
   const toggleControl = useStore.getState().toggleMapControls
   return (
