+++ features/geojson/Sidebar/Apps/index.tsx
@@ -1,10 +1,12 @@
 import { DarkMode } from "./DarkMode";
 import { GithubLink } from "./GithubLink";
+import { ControlOptions } from "./ControlOptions";
 
 export function AppsSettings() {
   return (
     <div>
       <DarkMode />
+      <ControlOptions />
       <GithubLink />
     </div>
   );
