+++ features/geojson/Sidebar/Layers/index.tsx
@@ -1,0 +1,9 @@
+import { MapLayer } from "./MapLayer";
+
+export const Layers = () => {
+  return (
+    <>
+      <MapLayer />
+    </>
+  );
+}
