+++ features/geojson/Sidebar/Legends/LegendList.tsx
@@ -29,7 +29,6 @@ import {
   createStyles,
   SegmentedControl,
 } from "@mantine/core";
-import { OptionWrapper } from "@components/Options";
